move QgsFieldModel and QgsFieldProxyModel to core and sipify
