+++ src/core/qgsfieldmodel.cpp
