+++ src/core/qgsfieldproxymodel.cpp
